+++ pylabrobot/resources/porvair/plates.py
@@ -7,7 +7,7 @@ from pylabrobot.resources.well import Well, WellBottomType, CrossSectionType
 from pylabrobot.resources.utils import create_equally_spaced_2d
 
 from pylabrobot.resources.volume_functions import calculate_liquid_volume_container_2segments_square_vbottom
-from pylabrobot.resources.height_functions import calculate_liquid_height_in_container_2segments_square_vbottom
+from pylabrobot.resources.height_volume_functions import calculate_liquid_height_in_container_2segments_square_vbottom
 
 # # # # # # # # # # Porvair_6_reservoir_47ml_Vb # # # # # # # # # #
