+++ pylabrobot/resources/thermo_fisher/plates.py
@@ -7,7 +7,7 @@ from pylabrobot.resources.utils import create_equally_spaced_2d
 from pylabrobot.resources.plate import Lid, Plate
 
 from pylabrobot.resources.volume_functions import calculate_liquid_volume_container_2segments_square_ubottom
-from pylabrobot.resources.height_functions import calculate_liquid_height_in_container_2segments_square_ubottom
+from pylabrobot.resources.height_volume_functions import calculate_liquid_height_in_container_2segments_square_ubottom
 
 
 # # # # # # # # # # ThermoScientific_96_DWP_1200ul_Rd # # # # # # # # # #
